Add an example Fly.io deployment of workers + autoscaler
The two worker TOML files define workers of different sizes only.
They're intended to be provided with two different queue IDs as Fly
secrets. The autoscaler is configured to target both of them by app name
prefix, and simply assumes that 1 worker can handle 1 job. It is not
configured to create or destroy any machines, so its up to the user to
define how many machines should exist for the app. The autoscaler will
start/stop them as needed.