Working autoscaler configuration for regional worker MIGs
The resource type filter needs to be applied or else the autoscaler will
assume the metric it's being configured with is a per-instance metric
and not per-group metric, and it will refuse to exist. We filter
resource type down to the one we want here, which is waaaaaay easier
than trying to figure out how to get the Google Cloud opentelemetry
processor/exporter to only submit per-group scaling compatible resource
types.