Properly scale each GCE instance group off its own set of metrics (not aggregated queue_length)
Toggle scalar data type so that the sent metric from otel is a INT64
type in Cloud Monitoring, instead of a DOUBLE

Provide a "nonexistent" default service for any requests to our ILB that
don't match host paths. This can be a simple 404 page/server later.

Switch to port 80 for http outside of the worker instances themselves

Include port 80 in host matches because opentelemetry is very strict
with its host header construction, and will flop on requests that curl
will handle gracefully

Don't include the metric descriptor as an explicit resource since
opentelemetry will create it the first time queue_length is submitted,
and the explicit resource gains no benefit over that